fix: solve problem compression with jsonl and gzip (http_to_data_lake)
`pandas.Dataframe.to_json` method does not compress the data when
passing a non-binary object as input. Example of log:

```bash
RuntimeWarning: compression has no effect when passing a non-binary object as input.
```

`HttpToDataLake` operator uploaded the file with `gzip` extension, but
the file was not compressed.

The solution was change `StringIO` by `BytesIO` for the memory file used
by pandas.