feat: metadata fields with '_*' convertion to datetime (FilesystemToDatabase)
When the operator encounters a metadata field that begins with the prefix '_*'
an attempt is made to convert to date automatically. If the parameter cannot
be converted to date the original value is left.

This is especially useful for date fields like `_DS` or `_INTERVAL_START`
where when the new column is added to the dataframe, it is added as if it
were a string instead of a date.