feat: add missing columns to database table or missing columns in source file
If the source file has columns not present in the database table, the operator
will add the missing columns to the database table and fill the previous rows of
those columns with null values.

If the database table has columns not present in the source file, the operator
will add the missing columns on source file with a null (`NaN`) value.